Minor Windigo API changes.

+++ dlg/dlgmain/DlgMainEvents.go
@@ -151,8 +151,8 @@ func (me *DlgMain) events() {
 				"Frees: %d",
 			nfo.LegalCopyright,
 			runtime.Version(),
-			stats.NumGC, wstr.FmtBytes(stats.HeapAlloc),
-			wstr.FmtBytes(stats.NextGC), stats.Frees)
+			stats.NumGC, wstr.FmtBytes(uint(stats.HeapAlloc)),
+			wstr.FmtBytes(uint(stats.NextGC)), stats.Frees)
 
 		ui.MsgOk(me.wnd, "About", caption, text)
 	})
