+++ id3v2/Body.go
@@ -104,7 +104,7 @@ func (me *BodyBinary) Clone() Body {
 }
 
 func (me *BodyBinary) AsText() string {
-	return wstr.FmtBytes(uint64(len(me.Bin)))
+	return wstr.FmtBytes(uint(len(me.Bin)))
 }
 
 func (me *BodyBinary) ForceText(text string) {
@@ -242,7 +242,7 @@ func (me *BodyPicture) Clone() Body {
 
 func (me *BodyPicture) AsText() string {
 	return fmt.Sprintf("%s %s %s",
-		PICNAMES[me.Type], me.Mime, wstr.FmtBytes(uint64(len(me.Bin))))
+		PICNAMES[me.Type], me.Mime, wstr.FmtBytes(uint(len(me.Bin))))
 }
 
 func (me *BodyPicture) ForceText(text string) {
